ref(ui): Restructure releases codebase
