+++ static/app/views/releases/detail/commitsAndFiles/commits.tsx
@@ -1,44 +1,44 @@
 import {Fragment} from 'react';
 import {RouteComponentProps} from 'react-router';
-import styled from '@emotion/styled';
 import {Location} from 'history';
 
-import FileChange from 'app/components/fileChange';
+import CommitRow from 'app/components/commitRow';
 import {Body, Main} from 'app/components/layouts/thirds';
 import LoadingIndicator from 'app/components/loadingIndicator';
 import Pagination from 'app/components/pagination';
 import {Panel, PanelBody, PanelHeader} from 'app/components/panels';
-import {t, tn} from 'app/locale';
-import {CommitFile, Organization, Project, Repository} from 'app/types';
+import {t} from 'app/locale';
+import {Commit, Organization, Project, Repository} from 'app/types';
 import {formatVersion} from 'app/utils/formatters';
 import routeTitleGen from 'app/utils/routeTitle';
 import AsyncView from 'app/views/asyncView';
 
+import {getCommitsByRepository, getQuery, getReposToRender} from '../utils';
+
 import EmptyState from './emptyState';
 import RepositorySwitcher from './repositorySwitcher';
-import {getFilesByRepository, getQuery, getReposToRender} from './utils';
 import withReleaseRepos from './withReleaseRepos';
 
 type Props = RouteComponentProps<{orgId: string; release: string}, {}> & {
   location: Location;
-  orgSlug: Organization['slug'];
   projectSlug: Project['slug'];
+  orgSlug: Organization['slug'];
   release: string;
   releaseRepos: Repository[];
   activeReleaseRepo?: Repository;
 } & AsyncView['props'];
 
 type State = {
-  fileList: CommitFile[];
+  commits: Commit[];
 } & AsyncView['state'];
 
-class FilesChanged extends AsyncView<Props, State> {
+class Commits extends AsyncView<Props, State> {
   getTitle() {
     const {params, projectSlug} = this.props;
     const {orgId} = params;
 
     return routeTitleGen(
-      t('Files Changed - Release %s', formatVersion(params.release)),
+      t('Commits - Release %s', formatVersion(params.release)),
       orgId,
       false,
       projectSlug
@@ -48,7 +48,7 @@ class FilesChanged extends AsyncView<Props, State> {
   getDefaultState(): State {
     return {
       ...super.getDefaultState(),
-      fileList: [],
+      commits: [],
     };
   }
 
@@ -61,14 +61,22 @@ class FilesChanged extends AsyncView<Props, State> {
   }
 
   getEndpoints(): ReturnType<AsyncView['getEndpoints']> {
-    const {activeReleaseRepo: activeRepository, location, release, orgSlug} = this.props;
+    const {
+      projectSlug,
+      activeReleaseRepo: activeRepository,
+      location,
+      orgSlug,
+      release,
+    } = this.props;
 
     const query = getQuery({location, activeRepository});
 
     return [
       [
-        'fileList',
-        `/organizations/${orgSlug}/releases/${encodeURIComponent(release)}/commitfiles/`,
+        'commits',
+        `/projects/${orgSlug}/${projectSlug}/releases/${encodeURIComponent(
+          release
+        )}/commits/`,
         {query},
       ],
     ];
@@ -79,63 +87,49 @@ class FilesChanged extends AsyncView<Props, State> {
   }
 
   renderContent() {
-    const {fileList, fileListPageLinks, loading} = this.state;
+    const {commits, commitsPageLinks, loading} = this.state;
     const {activeReleaseRepo} = this.props;
 
     if (loading) {
       return <LoadingIndicator />;
     }
 
-    if (!fileList.length) {
+    if (!commits.length) {
       return (
         <EmptyState>
           {!activeReleaseRepo
-            ? t('There are no changed files associated with this release.')
+            ? t('There are no commits associated with this release.')
             : t(
-                'There are no changed files associated with this release in the %s repository.',
+                'There are no commits associated with this release in the %s repository.',
                 activeReleaseRepo.name
               )}
         </EmptyState>
       );
     }
 
-    const filesByRepository = getFilesByRepository(fileList);
-    const reposToRender = getReposToRender(Object.keys(filesByRepository));
+    const commitsByRepository = getCommitsByRepository(commits);
+    const reposToRender = getReposToRender(Object.keys(commitsByRepository));
 
     return (
       <Fragment>
-        {reposToRender.map(repoName => {
-          const repoData = filesByRepository[repoName];
-          const files = Object.keys(repoData);
-          const fileCount = files.length;
-          return (
-            <Panel key={repoName}>
-              <PanelHeader>
-                <span>{repoName}</span>
-                <span>{tn('%s file changed', '%s files changed', fileCount)}</span>
-              </PanelHeader>
-              <PanelBody>
-                {files.map(filename => {
-                  const {authors} = repoData[filename];
-                  return (
-                    <StyledFileChange
-                      key={filename}
-                      filename={filename}
-                      authors={Object.values(authors)}
-                    />
-                  );
-                })}
-              </PanelBody>
-            </Panel>
-          );
-        })}
-        <Pagination pageLinks={fileListPageLinks} />
+        {reposToRender.map(repoName => (
+          <Panel key={repoName}>
+            <PanelHeader>{repoName}</PanelHeader>
+            <PanelBody>
+              {commitsByRepository[repoName]?.map(commit => (
+                <CommitRow key={commit.id} commit={commit} />
+              ))}
+            </PanelBody>
+          </Panel>
+        ))}
+        <Pagination pageLinks={commitsPageLinks} />
       </Fragment>
     );
   }
 
   renderBody() {
-    const {activeReleaseRepo, releaseRepos, router, location} = this.props;
+    const {location, router, activeReleaseRepo, releaseRepos} = this.props;
+
     return (
       <Fragment>
         {releaseRepos.length > 1 && (
@@ -160,15 +154,4 @@ class FilesChanged extends AsyncView<Props, State> {
   }
 }
 
-export default withReleaseRepos(FilesChanged);
-
-const StyledFileChange = styled(FileChange)`
-  border-radius: 0;
-  border-left: none;
-  border-right: none;
-  border-top: none;
-  :last-child {
-    border: none;
-    border-radius: 0;
-  }
-`;
+export default withReleaseRepos(Commits);
