+++ static/app/views/releases/detail/commitsAndFiles/filesChanged.tsx
@@ -1,43 +1,45 @@
 import {Fragment} from 'react';
 import {RouteComponentProps} from 'react-router';
+import styled from '@emotion/styled';
 import {Location} from 'history';
 
-import CommitRow from 'app/components/commitRow';
+import FileChange from 'app/components/fileChange';
 import {Body, Main} from 'app/components/layouts/thirds';
 import LoadingIndicator from 'app/components/loadingIndicator';
 import Pagination from 'app/components/pagination';
 import {Panel, PanelBody, PanelHeader} from 'app/components/panels';
-import {t} from 'app/locale';
-import {Commit, Organization, Project, Repository} from 'app/types';
+import {t, tn} from 'app/locale';
+import {CommitFile, Organization, Project, Repository} from 'app/types';
 import {formatVersion} from 'app/utils/formatters';
 import routeTitleGen from 'app/utils/routeTitle';
 import AsyncView from 'app/views/asyncView';
 
+import {getFilesByRepository, getQuery, getReposToRender} from '../utils';
+
 import EmptyState from './emptyState';
 import RepositorySwitcher from './repositorySwitcher';
-import {getCommitsByRepository, getQuery, getReposToRender} from './utils';
 import withReleaseRepos from './withReleaseRepos';
 
 type Props = RouteComponentProps<{orgId: string; release: string}, {}> & {
   location: Location;
-  projectSlug: Project['slug'];
   orgSlug: Organization['slug'];
+  projectSlug: Project['slug'];
   release: string;
   releaseRepos: Repository[];
   activeReleaseRepo?: Repository;
 } & AsyncView['props'];
 
 type State = {
-  commits: Commit[];
+  fileList: CommitFile[];
 } & AsyncView['state'];
 
-class Commits extends AsyncView<Props, State> {
+class FilesChanged extends AsyncView<Props, State> {
   getTitle() {
     const {params, projectSlug} = this.props;
     const {orgId} = params;
 
     return routeTitleGen(
-      t('Commits - Release %s', formatVersion(params.release)),
+      t('Files Changed - Release %s', formatVersion(params.release)),
       orgId,
       false,
       projectSlug
@@ -47,7 +49,7 @@ class Commits extends AsyncView<Props, State> {
   getDefaultState(): State {
     return {
       ...super.getDefaultState(),
-      commits: [],
+      fileList: [],
     };
   }
 
@@ -60,22 +62,14 @@ class Commits extends AsyncView<Props, State> {
   }
 
   getEndpoints(): ReturnType<AsyncView['getEndpoints']> {
-    const {
-      projectSlug,
-      activeReleaseRepo: activeRepository,
-      location,
-      orgSlug,
-      release,
-    } = this.props;
+    const {activeReleaseRepo: activeRepository, location, release, orgSlug} = this.props;
 
     const query = getQuery({location, activeRepository});
 
     return [
       [
-        'commits',
-        `/projects/${orgSlug}/${projectSlug}/releases/${encodeURIComponent(
-          release
-        )}/commits/`,
+        'fileList',
+        `/organizations/${orgSlug}/releases/${encodeURIComponent(release)}/commitfiles/`,
         {query},
       ],
     ];
@@ -86,49 +80,63 @@ class Commits extends AsyncView<Props, State> {
   }
 
   renderContent() {
-    const {commits, commitsPageLinks, loading} = this.state;
+    const {fileList, fileListPageLinks, loading} = this.state;
     const {activeReleaseRepo} = this.props;
 
     if (loading) {
       return <LoadingIndicator />;
     }
 
-    if (!commits.length) {
+    if (!fileList.length) {
       return (
         <EmptyState>
           {!activeReleaseRepo
-            ? t('There are no commits associated with this release.')
+            ? t('There are no changed files associated with this release.')
             : t(
-                'There are no commits associated with this release in the %s repository.',
+                'There are no changed files associated with this release in the %s repository.',
                 activeReleaseRepo.name
               )}
         </EmptyState>
       );
     }
 
-    const commitsByRepository = getCommitsByRepository(commits);
-    const reposToRender = getReposToRender(Object.keys(commitsByRepository));
+    const filesByRepository = getFilesByRepository(fileList);
+    const reposToRender = getReposToRender(Object.keys(filesByRepository));
 
     return (
       <Fragment>
-        {reposToRender.map(repoName => (
-          <Panel key={repoName}>
-            <PanelHeader>{repoName}</PanelHeader>
-            <PanelBody>
-              {commitsByRepository[repoName]?.map(commit => (
-                <CommitRow key={commit.id} commit={commit} />
-              ))}
-            </PanelBody>
-          </Panel>
-        ))}
-        <Pagination pageLinks={commitsPageLinks} />
+        {reposToRender.map(repoName => {
+          const repoData = filesByRepository[repoName];
+          const files = Object.keys(repoData);
+          const fileCount = files.length;
+          return (
+            <Panel key={repoName}>
+              <PanelHeader>
+                <span>{repoName}</span>
+                <span>{tn('%s file changed', '%s files changed', fileCount)}</span>
+              </PanelHeader>
+              <PanelBody>
+                {files.map(filename => {
+                  const {authors} = repoData[filename];
+                  return (
+                    <StyledFileChange
+                      key={filename}
+                      filename={filename}
+                      authors={Object.values(authors)}
+                    />
+                  );
+                })}
+              </PanelBody>
+            </Panel>
+          );
+        })}
+        <Pagination pageLinks={fileListPageLinks} />
       </Fragment>
     );
   }
 
   renderBody() {
-    const {location, router, activeReleaseRepo, releaseRepos} = this.props;
-
+    const {activeReleaseRepo, releaseRepos, router, location} = this.props;
     return (
       <Fragment>
         {releaseRepos.length > 1 && (
@@ -153,4 +161,15 @@ class Commits extends AsyncView<Props, State> {
   }
 }
 
-export default withReleaseRepos(Commits);
+export default withReleaseRepos(FilesChanged);
+
+const StyledFileChange = styled(FileChange)`
+  border-radius: 0;
+  border-left: none;
+  border-right: none;
+  border-top: none;
+  :last-child {
+    border: none;
+    border-radius: 0;
+  }
+`;
