+++ static/app/views/releases/list/index.tsx
@@ -20,7 +20,7 @@ import SearchBar from 'app/components/searchBar';
 import SmartSearchBar from 'app/components/smartSearchBar';
 import {DEFAULT_STATS_PERIOD, RELEASE_ADOPTION_STAGES} from 'app/constants';
 import {ALL_ACCESS_PROJECTS} from 'app/constants/globalSelectionHeader';
-import {desktop, mobile, PlatformKey, releaseHealth} from 'app/data/platformCategories';
+import {releaseHealth} from 'app/data/platformCategories';
 import {IconInfo} from 'app/icons';
 import {t} from 'app/locale';
 import ProjectsStore from 'app/stores/projectsStore';
@@ -43,15 +43,15 @@ import withProjects from 'app/utils/withProjects';
 import AsyncView from 'app/views/asyncView';
 
 import ReleaseArchivedNotice from '../detail/overview/releaseArchivedNotice';
+import {isMobileRelease} from '../utils';
 
-import ReleaseAdoptionChart from './releaseAdoptionChart';
 import ReleaseCard from './releaseCard';
-import ReleaseDisplayOptions from './releaseDisplayOptions';
-import ReleaseListRequest from './releaseListRequest';
-import ReleaseListSortOptions from './releaseListSortOptions';
-import ReleaseListStatusOptions from './releaseListStatusOptions';
-import ReleasePromo from './releasePromo';
-import {DisplayOption, SortOption, StatusOption} from './utils';
+import ReleasesAdoptionChart from './releasesAdoptionChart';
+import ReleasesDisplayOptions, {ReleasesDisplayOption} from './releasesDisplayOptions';
+import ReleasesPromo from './releasesPromo';
+import ReleasesRequest from './releasesRequest';
+import ReleasesSortOptions, {ReleasesSortOption} from './releasesSortOptions';
+import ReleasesStatusOptions, {ReleasesStatusOption} from './releasesStatusOptions';
 
 const supportedTags = {
   'release.version': {
@@ -77,9 +77,6 @@ const supportedTags = {
     name: 'release',
   },
 };
-
-export const isProjectMobileForReleases = (projectPlatform: PlatformKey) =>
-  ([...mobile, ...desktop] as string[]).includes(projectPlatform);
 
 type RouteParams = {
   orgId: string;
@@ -113,10 +110,10 @@ class ReleasesList extends AsyncView<Props, State> {
       ...pick(location.query, ['project', 'environment', 'cursor', 'query', 'sort']),
       summaryStatsPeriod: statsPeriod,
       per_page: 20,
-      flatten: activeSort === SortOption.DATE ? 0 : 1,
+      flatten: activeSort === ReleasesSortOption.DATE ? 0 : 1,
       adoptionStages: 1,
       status:
-        activeStatus === StatusOption.ARCHIVED
+        activeStatus === ReleasesStatusOption.ARCHIVED
           ? ReleaseStatus.Archived
           : ReleaseStatus.Active,
     };
@@ -156,42 +153,42 @@ class ReleasesList extends AsyncView<Props, State> {
     return typeof query === 'string' ? query : undefined;
   }
 
-  getSort(): SortOption {
+  getSort(): ReleasesSortOption {
     const {environments} = this.props.selection;
     const {sort} = this.props.location.query;
 
     // Require 1 environment for date adopted
-    if (sort === SortOption.ADOPTION && environments.length !== 1) {
-      return SortOption.DATE;
+    if (sort === ReleasesSortOption.ADOPTION && environments.length !== 1) {
+      return ReleasesSortOption.DATE;
     }
 
-    const sortExists = Object.values(SortOption).includes(sort);
+    const sortExists = Object.values(ReleasesSortOption).includes(sort);
     if (sortExists) {
       return sort;
     }
 
-    return SortOption.DATE;
+    return ReleasesSortOption.DATE;
   }
 
-  getDisplay(): DisplayOption {
+  getDisplay(): ReleasesDisplayOption {
     const {display} = this.props.location.query;
 
     switch (display) {
-      case DisplayOption.USERS:
-        return DisplayOption.USERS;
+      case ReleasesDisplayOption.USERS:
+        return ReleasesDisplayOption.USERS;
       default:
-        return DisplayOption.SESSIONS;
+        return ReleasesDisplayOption.SESSIONS;
     }
   }
 
-  getStatus(): StatusOption {
+  getStatus(): ReleasesStatusOption {
     const {status} = this.props.location.query;
 
     switch (status) {
-      case StatusOption.ARCHIVED:
-        return StatusOption.ARCHIVED;
+      case ReleasesStatusOption.ARCHIVED:
+        return ReleasesStatusOption.ARCHIVED;
       default:
-        return StatusOption.ACTIVE;
+        return ReleasesStatusOption.ACTIVE;
     }
   }
 
@@ -229,14 +226,26 @@ class ReleasesList extends AsyncView<Props, State> {
     const {location, router} = this.props;
 
     let sort = location.query.sort;
-    if (sort === SortOption.USERS_24_HOURS && display === DisplayOption.SESSIONS)
-      sort = SortOption.SESSIONS_24_HOURS;
-    else if (sort === SortOption.SESSIONS_24_HOURS && display === DisplayOption.USERS)
-      sort = SortOption.USERS_24_HOURS;
-    else if (sort === SortOption.CRASH_FREE_USERS && display === DisplayOption.SESSIONS)
-      sort = SortOption.CRASH_FREE_SESSIONS;
-    else if (sort === SortOption.CRASH_FREE_SESSIONS && display === DisplayOption.USERS)
-      sort = SortOption.CRASH_FREE_USERS;
+    if (
+      sort === ReleasesSortOption.USERS_24_HOURS &&
+      display === ReleasesDisplayOption.SESSIONS
+    )
+      sort = ReleasesSortOption.SESSIONS_24_HOURS;
+    else if (
+      sort === ReleasesSortOption.SESSIONS_24_HOURS &&
+      display === ReleasesDisplayOption.USERS
+    )
+      sort = ReleasesSortOption.USERS_24_HOURS;
+    else if (
+      sort === ReleasesSortOption.CRASH_FREE_USERS &&
+      display === ReleasesDisplayOption.SESSIONS
+    )
+      sort = ReleasesSortOption.CRASH_FREE_SESSIONS;
+    else if (
+      sort === ReleasesSortOption.CRASH_FREE_SESSIONS &&
+      display === ReleasesDisplayOption.USERS
+    )
+      sort = ReleasesSortOption.CRASH_FREE_USERS;
 
     router.push({
       ...location,
@@ -314,7 +323,7 @@ class ReleasesList extends AsyncView<Props, State> {
       );
     }
 
-    if (activeSort === SortOption.USERS_24_HOURS) {
+    if (activeSort === ReleasesSortOption.USERS_24_HOURS) {
       return (
         <EmptyStateWarning small>
           {t('There are no releases with active user data (users in the last 24 hours).')}
@@ -322,7 +331,7 @@ class ReleasesList extends AsyncView<Props, State> {
       );
     }
 
-    if (activeSort === SortOption.SESSIONS_24_HOURS) {
+    if (activeSort === ReleasesSortOption.SESSIONS_24_HOURS) {
       return (
         <EmptyStateWarning small>
           {t(
@@ -332,7 +341,10 @@ class ReleasesList extends AsyncView<Props, State> {
       );
     }
 
-    if (activeSort === SortOption.BUILD || activeSort === SortOption.SEMVER) {
+    if (
+      activeSort === ReleasesSortOption.BUILD ||
+      activeSort === ReleasesSortOption.SEMVER
+    ) {
       return (
         <EmptyStateWarning small>
           {t('There are no releases with semantic versioning.')}
@@ -340,7 +352,7 @@ class ReleasesList extends AsyncView<Props, State> {
       );
     }
 
-    if (activeSort !== SortOption.DATE) {
+    if (activeSort !== ReleasesSortOption.DATE) {
       const relativePeriod = getRelativeSummary(
         statsPeriod || DEFAULT_STATS_PERIOD
       ).toLowerCase();
@@ -352,7 +364,7 @@ class ReleasesList extends AsyncView<Props, State> {
       );
     }
 
-    if (activeStatus === StatusOption.ARCHIVED) {
+    if (activeStatus === ReleasesStatusOption.ARCHIVED) {
       return (
         <EmptyStateWarning small>
           {t('There are no archived releases.')}
@@ -361,7 +373,7 @@ class ReleasesList extends AsyncView<Props, State> {
     }
 
     return (
-      <ReleasePromo
+      <ReleasesPromo
         organization={organization}
         projectId={selection.projects.filter(p => p !== ALL_ACCESS_PROJECTS)[0]}
       />
@@ -411,7 +423,10 @@ class ReleasesList extends AsyncView<Props, State> {
     );
   }
 
-  renderInnerBody(activeDisplay: DisplayOption, showReleaseAdoptionStages: boolean) {
+  renderInnerBody(
+    activeDisplay: ReleasesDisplayOption,
+    showReleaseAdoptionStages: boolean
+  ) {
     const {location, selection, organization, router} = this.props;
     const {releases, reloading, releasesPageLinks} = this.state;
 
@@ -424,7 +439,7 @@ class ReleasesList extends AsyncView<Props, State> {
     }
 
     return (
-      <ReleaseListRequest
+      <ReleasesRequest
         releases={releases.map(({version}) => version)}
         organization={organization}
         selection={selection}
@@ -439,14 +454,13 @@ class ReleasesList extends AsyncView<Props, State> {
             selection.projects[0] !== ALL_ACCESS_PROJECTS;
           const selectedProject = this.getSelectedProject();
           const isMobileProject =
-            selectedProject?.platform &&
-            isProjectMobileForReleases(selectedProject.platform);
+            selectedProject?.platform && isMobileRelease(selectedProject.platform);
 
           return (
             <Fragment>
               {singleProjectSelected && this.projectHasSessions && isMobileProject && (
                 <Feature features={['organizations:release-adoption-chart']}>
-                  <ReleaseAdoptionChart
+                  <ReleasesAdoptionChart
                     organization={organization}
                     selection={selection}
                     location={location}
@@ -475,7 +489,7 @@ class ReleasesList extends AsyncView<Props, State> {
             </Fragment>
           );
         }}
-      </ReleaseListRequest>
+      </ReleasesRequest>
     );
   }
 
@@ -492,7 +506,7 @@ class ReleasesList extends AsyncView<Props, State> {
     const hasAnyMobileProject = selection.projects
       .map(id => `${id}`)
       .map(ProjectsStore.getById)
-      .some(project => project?.platform && isProjectMobileForReleases(project.platform));
+      .some(project => project?.platform && isMobileRelease(project.platform));
     const showReleaseAdoptionStages =
       hasReleaseStages && hasAnyMobileProject && selection.environments.length === 1;
     const hasReleasesSetup = releases && releases.length > 0;
@@ -544,18 +558,18 @@ class ReleasesList extends AsyncView<Props, State> {
                 />
               )}
               <DropdownsWrapper>
-                <ReleaseListStatusOptions
+                <ReleasesStatusOptions
                   selected={activeStatus}
                   onSelect={this.handleStatus}
                 />
-                <ReleaseListSortOptions
+                <ReleasesSortOptions
                   selected={activeSort}
                   selectedDisplay={activeDisplay}
                   onSelect={this.handleSortBy}
                   environments={selection.environments}
                   organization={organization}
                 />
-                <ReleaseDisplayOptions
+                <ReleasesDisplayOptions
                   selected={activeDisplay}
                   onSelect={this.handleDisplay}
                 />
@@ -563,7 +577,7 @@ class ReleasesList extends AsyncView<Props, State> {
             </SortAndFilterWrapper>
 
             {!reloading &&
-              activeStatus === StatusOption.ARCHIVED &&
+              activeStatus === ReleasesStatusOption.ARCHIVED &&
               !!releases?.length && <ReleaseArchivedNotice multi />}
 
             {error
