--- static/app/views/releases/list/utils.tsx
@@ -1,21 +1,0 @@
-export enum SortOption {
-  CRASH_FREE_USERS = 'crash_free_users',
-  CRASH_FREE_SESSIONS = 'crash_free_sessions',
-  USERS_24_HOURS = 'users_24h',
-  SESSIONS_24_HOURS = 'sessions_24h',
-  SESSIONS = 'sessions',
-  DATE = 'date',
-  BUILD = 'build',
-  SEMVER = 'semver',
-  ADOPTION = 'adoption',
-}
-
-export enum DisplayOption {
-  USERS = 'users',
-  SESSIONS = 'sessions',
-}
-
-export enum StatusOption {
-  ACTIVE = 'active',
-  ARCHIVED = 'archived',
-}
