+++ static/app/views/releases/utils/index.tsx
@@ -4,14 +4,15 @@ import round from 'lodash/round';
 import moment from 'moment';
 
 import {DateTimeObject} from 'app/components/charts/utils';
+import ExternalLink from 'app/components/links/externalLink';
 import {getParams} from 'app/components/organizations/globalSelectionHeader/getParams';
 import {PAGE_URL_PARAM, URL_PARAM} from 'app/constants/globalSelectionHeader';
-import {tn} from 'app/locale';
+import {desktop, mobile, PlatformKey} from 'app/data/platformCategories';
+import {t, tct} from 'app/locale';
 import {Release, ReleaseStatus} from 'app/types';
+import {Theme} from 'app/utils/theme';
 import {MutableSearch} from 'app/utils/tokenizeSearch';
 import {IssueSortOptions} from 'app/views/issueList/utils';
-
-import {DisplayOption} from '../list/utils';
 
 export const CRASH_FREE_DECIMAL_THRESHOLD = 95;
 
@@ -63,17 +64,6 @@ export const getSessionStatusPercent = (percent: number, absolute = true) => {
 export const displaySessionStatusPercent = (percent: number, absolute = true) => {
   return `${getSessionStatusPercent(percent, absolute).toLocaleString()}\u0025`;
 };
-
-export const displayCrashFreeDiff = (
-  diffPercent: number,
-  crashFreePercent?: number | null
-) =>
-  `${Math.abs(
-    round(
-      diffPercent,
-      crashFreePercent && crashFreePercent > CRASH_FREE_DECIMAL_THRESHOLD ? 3 : 0
-    )
-  ).toLocaleString()}\u0025`;
 
 export const getReleaseNewIssuesUrl = (
   orgSlug: string,
@@ -136,14 +126,6 @@ export const getReleaseHandledIssuesUrl = (
 
 export const isReleaseArchived = (release: Release) =>
   release.status === ReleaseStatus.Archived;
-
-export function releaseDisplayLabel(displayOption: DisplayOption, count?: number | null) {
-  if (displayOption === DisplayOption.USERS) {
-    return tn('user', 'users', count);
-  }
-
-  return tn('session', 'sessions', count);
-}
 
 export type ReleaseBounds = {releaseStart?: string | null; releaseEnd?: string | null};
 
@@ -211,3 +193,40 @@ export function getReleaseParams({location, releaseBounds}: GetReleaseParams) {
 
   return params;
 }
+
+const adoptionStagesLink = (
+  <ExternalLink href="https://docs.sentry.io/product/releases/health/#adoption-stages" />
+);
+
+export const ADOPTION_STAGE_LABELS: Record<
+  string,
+  {name: string; tooltipTitle: JSX.Element; type: keyof Theme['tag']}
+> = {
+  low_adoption: {
+    name: t('Low Adoption'),
+    tooltipTitle: tct(
+      'This release has a low percentage of sessions compared to other releases in this project. [link:Learn more]',
+      {link: adoptionStagesLink}
+    ),
+    type: 'warning',
+  },
+  adopted: {
+    name: t('Adopted'),
+    tooltipTitle: tct(
+      'This release has a high percentage of sessions compared to other releases in this project. [link:Learn more]',
+      {link: adoptionStagesLink}
+    ),
+    type: 'success',
+  },
+  replaced: {
+    name: t('Replaced'),
+    tooltipTitle: tct(
+      'This release was previously Adopted, but now has a lower level of sessions compared to other releases in this project. [link:Learn more]',
+      {link: adoptionStagesLink}
+    ),
+    type: 'default',
+  },
+};
+
+export const isMobileRelease = (releaseProjectPlatform: PlatformKey) =>
+  ([...mobile, ...desktop] as string[]).includes(releaseProjectPlatform);
