+++ tests/js/spec/views/releases/detail/releaseActions.spec.jsx
@@ -4,7 +4,7 @@ import {mountWithTheme} from 'sentry-test/enzyme';
 import {initializeOrg} from 'sentry-test/initializeOrg';
 import {mountGlobalModal} from 'sentry-test/modal';
 
-import ReleaseActions from 'app/views/releases/detail/releaseActions';
+import ReleaseActions from 'app/views/releases/detail/header/releaseActions';
 
 describe('ReleaseActions', function () {
   const {organization} = initializeOrg();
